pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flight-deal-agent"
version = "0.2.0"
description = "按区域监控往返特价航班的 agentic 后端"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "Jeffrey Deng" }]
keywords = ["flights", "travel", "agent", "monitoring"]
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.5",
    "apscheduler>=3.10,<4",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
]

[project.scripts]
flight-deal-agent = "flight_deal_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["flight_deal_agent"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
